+++ src/rcrypto.h
@@ -1,0 +1,1 @@
+void rcrypto(void *buf, size_t n);