+++ tests/testthat/test-argon.R
@@ -1,10 +1,10 @@
 
 test_that("argon2 works", {
 
-  salt <- isaac(16)
+  salt <- rcrypto(16)
   res1 <- argon2("my secret", type = 'raw', salt = salt)
   res2 <- argon2("my secret", type = 'raw', salt = salt)
-  res3 <- argon2("my secret", type = 'raw', salt = isaac(16))
+  res3 <- argon2("my secret", type = 'raw', salt = rcrypto(16))
 
   expect_true(is.raw(res1))
   expect_identical(res1, res2)
@@ -16,10 +16,10 @@ test_that("argon2 works", {
 
 test_that("argon2 works", {
   
-  salt <- isaac(16)
+  salt <- rcrypto(16)
   res1 <- argon2("my secret", type = 'string', salt = salt)
   res2 <- argon2("my secret", type = 'string', salt = salt)
-  res3 <- argon2("my secret", type = 'string', salt = isaac(16))
+  res3 <- argon2("my secret", type = 'string', salt = rcrypto(16))
   
   expect_true(is.character(res1))
   expect_identical(res1, res2)