+++ NEWS.md
@@ -1,4 +1,8 @@
 
+# rmonocypher 0.1.2  2024-04-25
+
+* Replace the ISAAC CSPRNG with the OS-provided, platform specific CSPRNG
+
 # rmonocypher 0.1.1  2024-04-22
 
 * Added x25519 key exchange functions `create_public_key()` and